% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dotproduct}
\alias{dotproduct}
\title{Computes the dot product of two iterable objects}
\usage{
dotproduct(vec1, vec2)
}
\arguments{
\item{vec1}{the first}

\item{vec2}{the second iterable object}
}
\value{
the dot product of the iterators
}
\description{
Returns the dot product of two numeric iterables of equal length
}
\examples{
it <- iterators::iter(1:3)
it2 <- iterators::iter(4:6)
dotproduct(it, it2) # 32

it <- iterators::iter(1:4)
it2 <- iterators::iter(7:10)
dotproduct(1:4, 7:10) # 90
}

