% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/icombinations.r
\name{icombinations}
\alias{icombinations}
\title{Iterator that generates all combinations of a vector taken m at a time.}
\usage{
icombinations(object, m)
}
\arguments{
\item{object}{vector}
}
\value{
iterator that generates all combinations of \code{object}
}
\description{
Constructs an iterator generates all combinations of a vector taken \code{m}
at a time. This function is similar to \code{\link[utils]{combn}}.
}
\details{
The implementation is loosely based on that of Python's itertools.
}
\examples{
it <- icombinations(1:4, m=2)

iterators::nextElem(it) # c(1, 2)
iterators::nextElem(it) # c(1, 3)
iterators::nextElem(it) # c(1, 4)
iterators::nextElem(it) # c(2, 3)
iterators::nextElem(it) # c(2, 4)
iterators::nextElem(it) # c(3, 4)

it2 <- icombinations(1:5, m=2)
as.list(it2)
utils::combn(x=1:5, m=2, simplify=FALSE)
}

