% Generated by roxygen2 (4.0.1): do not edit by hand
\name{icompress}
\alias{icompress}
\title{Iterator that filters elements where corresponding selector is false.}
\usage{
icompress(object, selectors)
}
\arguments{
\item{object}{an iterable object}

\item{selectors}{an iterable that determines whether the corresponding
element in \code{object} is returned.}
}
\value{
iterator object
}
\description{
Constructs an iterator that filters elements from iterable returning only
those for which the corresponding element from \code{selectors} is
\code{TRUE}.
}
\details{
The iterator stops when either \code{object} or \code{selectors} has been
exhausted.
}
\examples{
# Filters out odd numbers and retains only even numbers
n <- 10
selectors <- rep(c(F, T), n)
it <- icompress(seq_len(n), selectors)
nextElem(it) # 2
nextElem(it) # 4
nextElem(it) # 6
nextElem(it) # 8
nextElem(it) # 10

# Similar idea here but anonymous function is used to filter out even
# numbers
n <- 10
it2 <- icompress(seq_len(10), rep(c(T, F), n))
nextElem(it) # 1
nextElem(it) # 3
nextElem(it) # 5
nextElem(it) # 7
nextElem(it) # 9

it3 <- icompress(letters, letters \%in\% c('a', 'e', 'i', 'o', 'u'))
nextElem(it) # a
nextElem(it) # e
nextElem(it) # i
nextElem(it) # o
nextElem(it) # u
}

