% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ienumerate}
\alias{ienum}
\alias{ienumerate}
\title{Iterator that returns the elements of an object along with their indices}
\usage{
ienumerate(object)

ienum(object)
}
\arguments{
\item{object}{object to return indefinitely.}
}
\value{
iterator that returns the values of \code{object} along with the
index of the object.
}
\description{
Constructs an iterator that returns the elements of an object along with each
element's indices. Enumeration is useful when looping through an
\code{object} and a counter is required.
}
\details{
This function is intended to follow the convention used in Python's
\code{enumerate} function where the primary difference is that a list is
returned instead of Python's \code{tuple} construct.

Each call to \code{\link[iterators]{nextElem}} returns a list with two
elements:
\describe{
  \item{index:}{a counter}
  \item{value:}{the current value of \code{object}}
}

\code{ienum} is an alias to \code{ienumerate} to save a few keystrokes.
}
\examples{
set.seed(42)
it <- ienumerate(rnorm(5))
as.list(it)

# Iterates through the columns of the iris data.frame
it2 <- ienum(iris)
iterators::nextElem(it2)
iterators::nextElem(it2)
iterators::nextElem(it2)
iterators::nextElem(it2)
iterators::nextElem(it2)
}

