% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ilength}
\alias{ilength}
\title{Consumes an iterator and computes its length}
\usage{
ilength(object)
}
\arguments{
\item{object}{an iterable object}
}
\value{
the number of elements in the iterator
}
\description{
Counts the number of elements in an iterator. NOTE: The iterator is consumed
in the process.
}
\examples{
ilength(1:5) == length(1:5)

it <- iterators::iter(1:5)
ilength(it) == length(1:5)

it2 <- ichain(1:3, 4:5, 6)
ilength(it2)

it3 <- ichain(1:3, levels(iris$Species))
ilength(it3)
}

