% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipairwise}
\alias{ipairwise}
\title{Iterator that returns elements of an object in pairs}
\usage{
ipairwise(object)
}
\arguments{
\item{object}{an iterable object}
}
\value{
an iterator that returns pairwise elements
}
\description{
Constructs an iterator of an iterable \code{object} that returns its elements
in pairs.
}
\examples{
it <- ipairwise(iterators::iter(letters[1:4]))
iterators::nextElem(it) # list("a", "b")
iterators::nextElem(it) # list("b", "c")
iterators::nextElem(it) # list("c", "d")

it2 <- ipairwise(1:5)
iterators::nextElem(it2) # list(1, 2)
iterators::nextElem(it2) # list(2, 3)
iterators::nextElem(it2) # list(3, 4)
iterators::nextElem(it2) # list(4, 5)
}

