% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iproduct}
\alias{iproduct}
\title{Iterator that returns the Cartesian product of the arguments.}
\usage{
iproduct(...)
}
\arguments{
\item{...}{multiple arguments}
}
\value{
iterator that iterates through each element from the Cartesian
product
}
\description{
Constructs an iterator that is the Cartesian product of each of the arguments.
}
\details{
Although they share the same end goal, \code{iproduct} can yield drastic
memory savings compared to \code{\link[base]{expand.grid}}.
}
\examples{
it <- iproduct(1:3, 4:5)
nextElem(it) # list(1, 4)
nextElem(it) # list(1, 5)
nextElem(it) # list(2, 4)
nextElem(it) # list(2, 5)
nextElem(it) # list(3, 4)
nextElem(it) # list(3, 5)

it2 <- iproduct(a=1:2, b=3:4, c=5:6)
nextElem(it2) # list(a=1, b=3, c=5)
nextElem(it2) # list(a=1, b=3, c=6)
nextElem(it2) # list(a=1, b=4, c=5)
nextElem(it2) # list(a=1, b=4, c=6)
nextElem(it2) # list(a=2, b=3, c=5)
nextElem(it2) # list(a=2, b=3, c=6)
nextElem(it2) # list(a=2, b=4, c=5)
nextElem(it2) # list(a=2, b=4, c=6)
}

