% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_iterator}
\alias{is_iterator}
\title{Helper function that determines whether is an iterator object}
\usage{
is_iterator(object)
}
\arguments{
\item{object}{an R object}
}
\value{
logical value indicating whether \code{object} is of class
\code{iter}
}
\description{
Returns \code{TRUE} if the \code{object} is an object of class \code{iter},
and \code{FALSE} otherwise.
}

