% Generated by roxygen2 (4.0.1): do not edit by hand
\name{itee}
\alias{itee}
\title{Returns a list of n independent iterators from a single iterable object}
\usage{
itee(object, n = 2)
}
\arguments{
\item{object}{an iterable object}

\item{n}{the number of iterables to return}
}
\value{
a list of \code{n} iterators
}
\description{
Constructs a list of \code{n} iterators, each of which iterates through an
iterable \code{object}.
}
\examples{
iter_list <- itee(1:5, n=3)
nextElem(iter_list[[1]]) # 1
nextElem(iter_list[[1]]) # 2
nextElem(iter_list[[1]]) # 3

nextElem(iter_list[[2]]) # 1
nextElem(iter_list[[2]]) # 2

nextElem(iter_list[[3]]) # 1
nextElem(iter_list[[3]]) # 2

nextElem(iter_list[[1]]) # 4
nextElem(iter_list[[1]]) # 5

nextElem(iter_list[[2]]) # 3
nextElem(iter_list[[2]]) # 4
nextElem(iter_list[[2]]) # 5

nextElem(iter_list[[3]]) # 3
nextElem(iter_list[[3]]) # 4
nextElem(iter_list[[3]]) # 5
}

