% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iter_length}
\alias{iter_length}
\title{Helper function that determines the length of an iterator object}
\usage{
iter_length(object, default = 1)
}
\arguments{
\item{object}{an iterator object}

\item{default}{the value returned when an iterator has \code{NULL} length}
}
\value{
integer
}
\description{
Returns the length of an iterator object. In the case that the iterator's
length is \code{NULL}, a value of 1 is returned by default. This value can be
set using the \code{default} argument.
}

