% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{itertools2}
\alias{itertools2}
\alias{itertools2-package}
\alias{package-itertools2}
\title{itertools2: Functions creating iterators for efficient looping}
\description{
The R package \code{itertools2} is a port of Python's excellent itertools
module \url{https://docs.python.org/2/library/itertools.html} to R for
efficient looping and is a replacement for the existing itertools R package
\url{https://r-forge.r-project.org/projects/itertools/}.
}

