% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itripletwise.r
\name{itripletwise}
\alias{itripletwise}
\title{Iterator that returns elements of an object in triplets}
\usage{
itripletwise(object)
}
\arguments{
\item{object}{an iterable object}
}
\value{
an iterator that returns tripletwise elements
}
\description{
Constructs an iterator of an iterable \code{object} that returns its elements
in pairs.
}
\examples{
it <- itripletwise(iterators::iter(letters[1:4]))
iterators::nextElem(it) # list("a", "b", "c")
iterators::nextElem(it) # list("b", "c", "d")

it2 <- itripletwise(1:5)
iterators::nextElem(it2) # list(1, 2, 3)
iterators::nextElem(it2) # list(2, 3, 4)
iterators::nextElem(it2) # list(3, 4, 5)
}

