% Generated by roxygen2 (4.0.1): do not edit by hand
\name{izip_longest}
\alias{izip_longest}
\title{Iterator that iterates through several iterables concurrently.}
\usage{
izip_longest(..., fill = NA)
}
\arguments{
\item{...}{multiple arguments to iterate through in sequence}

\item{fill}{the value used to replace missing values when the iterables in
\code{...} are of uneven length}
}
\value{
iterator that iterates through each argument in sequence
}
\description{
The resulting iterator aggregates elements from each of the iterables into a
list from each iteration. Used for lock-step iteration over several iterables
at a time.
}
\details{
Although similar to \code{\link[itertools2]{izip}}, missing values are
replaced with \code{fill} if the iterables are of uneven length, and
Iteration continues until the longest iterable is exhausted.
}
\examples{
it <- izip(1:3, 4:6, 7:9)
nextElem(it) # list(1, 4, 7)
nextElem(it) # list(2, 5, 8)
nextElem(it) # list(3, 6, 9)

it2 <- izip(1:3, 4:8)
nextElem(it2) # list(1, 4)
nextElem(it2) # list(2, 5)
nextElem(it2) # list(3, 6)
nextElem(it2) # list(NA, 7)
nextElem(it2) # list(NA, 8)

it3 <- izip(1:3, 4:7, levels(iris$Species), fill="w00t")
nextElem(it3) # list(1, 4, "setosa")
nextElem(it3) # list(2, 5, "versicolor")
nextElem(it3) # list(3, 6, "virginica")
nextElem(it3) # list("w00t", 7, "w00t")
}

