% Generated by roxygen2 (4.0.1): do not edit by hand
\name{take}
\alias{take}
\title{Return the first n elements of an iterable object as a list}
\usage{
take(object, n = 1)
}
\arguments{
\item{object}{an iterable object}

\item{n}{the number of elements to return in the list}
}
\value{
a list of the first \code{n} items of the iterable \code{object}
}
\description{
Returns the first \code{n} elements of an iterable \code{object} as a list.
If \code{n} is larger than the number of elements in \code{object}, the
entire iterator is consumed.
}
\examples{
take(iterators::iter(1:10), 3) # 1 2 3

take(iterators::iter(1:5), 10) # 1 2 3 4 5
}

