% Generated by roxygen2 (4.0.1): do not edit by hand
\name{try_nextElem}
\alias{try_nextElem}
\title{Calls iterators::nextElem(). If error, returns default value.}
\usage{
try_nextElem(object, default = NA, silent = TRUE)
}
\arguments{
\item{object}{an iterable object}

\item{default}{default value returned if a StopIteration exception is thrown}

\item{silent}{Should any errors be suppressed without explicitly notifying
the user? Default. Yes}
}
\value{
the next element of \code{object}
}
\description{
Returns the next element of \code{object}. In the case a StopIteration
exception is thrown, the \code{default} value is returned instead.
}

