% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ichain}
\alias{ichain}
\title{Iterator that chains multiple arguments together into a single iterator}
\usage{
ichain(...)
}
\arguments{
\item{...}{multiple arguments to iterate through in sequence}
}
\value{
iterator that iterates through each argument in sequence
}
\description{
Generates an iterator that returns elements from the first argument until it
is exhausted. Then generates an iterator from the next argument and returns
elements from it. This process continues until all arguments are exhausted
Chaining is useful for treating consecutive sequences as a single sequence.
}
\examples{
it <- ichain(1:3, 4:5, 6)
as.list(it)

it2 <- ichain(1:3, levels(iris$Species))
as.list(it2)
}

