% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stop_iteration}
\alias{stop_iteration}
\title{Helper function that determines whether an object inherits from a
StopIteration exception}
\usage{
stop_iteration(object)
}
\arguments{
\item{object}{an R object}
}
\value{
\code{TRUE} if \code{object} resulted from a StopIteration
exception. Otherwise, \code{FALSE}.
}
\description{
Returns \code{TRUE} if the \code{object} resulted from a StopIteration
exception when \code{\link[iterators]{nextElem}} is called, and \code{FALSE}
otherwise.
}

